backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis[hiredis]==5.0.1
httpx==0.27.0
python-multipart==0.0.9
apscheduler==3.10.4
qrcode[pil]==7.4.2
firebase-admin==6.4.0
twilio==9.0.0
